Global replace of single-precision floats with CGFloat casts. (#5718)
Replacing all non-integral, single-precision floating point literals with double-precision values explicitly cast to CGFloat.  For example, `0.1f` should become `(CGFloat)0.1`.

Regex used:
```
([0-9]*\.)([0-9]*?)([1-9]+)([0-9]*?)[fF]
```

PiperOrigin-RevId: 220683126

+++ components/ActionSheet/tests/unit/MDCActionSheetTestHelper.m
@@ -22,13 +22,19 @@
 @implementation MDCActionSheetTestHelper
 
 + (NSArray *)colorsToTest {
-  UIColor *rgbColor = [UIColor colorWithRed:0.7f green:0.7f blue:0.7f alpha:0.7f];
-  UIColor *hsbColor = [UIColor colorWithHue:0.8f saturation:0.8f brightness:0.8f alpha:0.8f];
-  UIColor *blackWithAlpha = [UIColor.greenColor colorWithAlphaComponent:0.8f];
+  UIColor *rgbColor = [UIColor colorWithRed:(CGFloat)0.7
+                                      green:(CGFloat)0.7
+                                       blue:(CGFloat)0.7
+                                      alpha:(CGFloat)0.7];
+  UIColor *hsbColor = [UIColor colorWithHue:(CGFloat)0.8
+                                 saturation:(CGFloat)0.8
+                                 brightness:(CGFloat)0.8
+                                      alpha:(CGFloat)0.8];
+  UIColor *blackWithAlpha = [UIColor.greenColor colorWithAlphaComponent:(CGFloat)0.8];
   UIColor *black = UIColor.blackColor;
   CIColor *ciColor = [[CIColor alloc] initWithColor:UIColor.blackColor];
   UIColor *uiColorFromCIColor = [UIColor colorWithCIColor:ciColor];
-  UIColor *whiteColor = [UIColor colorWithWhite:0.5f alpha:0.5f];
+  UIColor *whiteColor = [UIColor colorWithWhite:(CGFloat)0.5 alpha:(CGFloat)0.5];
   return @[ rgbColor, hsbColor, blackWithAlpha, black, uiColorFromCIColor, whiteColor ];
 }
